Refactor PBT and extract stack and device classes
